feat(webpack): Improve webpack support
This PR improves webpack support by adding a choice for webpack/browserify when first running the generator.

If webpack is chosen, stylesheets will be bundled using native webpack loaders, and files will be copied to the output folder similarly.

It also changes the bootstrap dep to the sass version so that the font path can be overriden.

My preliminary tests seem ok, let me know if you find anything wrong.